[REF] Change sorting to give a consistent behaviour
The individual sort toggles were changed to a SortToggle component.
This can potentially be used in the future for new sort methods.
Main's state.filters.sortBy was changed from a string to an array to
allow for multiple sorting methods. A bug was fixed where the sorting
methods were altering the state.cards directly instead of making a new
array and updating state.cardsToShow.